moved SimpleAccountCard to composables/card package
