+++ app/src/main/java/com/babylon/wallet/android/presentation/dapp/authorized/verifyentities/VerifyEntitiesContent.kt
@@ -31,7 +31,7 @@ import com.babylon.wallet.android.presentation.ui.RadixWalletPreviewTheme
 import com.babylon.wallet.android.presentation.ui.composables.BackIconType
 import com.babylon.wallet.android.presentation.ui.composables.RadixBottomBar
 import com.babylon.wallet.android.presentation.ui.composables.RadixCenteredTopAppBar
-import com.babylon.wallet.android.presentation.ui.composables.SimpleAccountCardWithAddress
+import com.babylon.wallet.android.presentation.ui.composables.card.SimpleAccountCardWithAddress
 import com.babylon.wallet.android.presentation.ui.composables.Thumbnail
 import com.babylon.wallet.android.presentation.ui.composables.card.SimplePersonaCard
 import com.babylon.wallet.android.presentation.ui.composables.displayName
